refactor: move withLogging middleware to server/middleware/logging.ts
